// convex/_generated/server.d.ts
export declare const query: QueryBuilder<DataModel, "public">;
// ...
export declare const mutation: MutationBuilder<DataModel, "public">;
// ...
export declare const action: ActionBuilder<DataModel, "public">;
// ...
export declare const app: {
  sharded_counter: {
    counter: {
      get: FunctionReference<"query", "internal", { key: string }, any>;
      increment: FunctionReference<
        "mutation",
        "internal",
        { key: string },
        any
      >;
    };
  };
};